Refactored SelfManager, moved registration to ZeroInstallInstance

+++ src/Commands/Desktop/SelfManagement/SelfManager.DesktopIntegration.cs
@@ -2,6 +2,8 @@
 // Licensed under the GNU Lesser Public License
 
 using System.IO;
+using NanoByte.Common.Info;
+using NanoByte.Common.Native;
 using NanoByte.Common.Tasks;
 using ZeroInstall.Commands.Properties;
 using ZeroInstall.DesktopIntegration.Windows;
@@ -10,19 +12,36 @@ namespace ZeroInstall.Commands.Desktop.SelfManagement
 {
     partial class SelfManager
     {
+        private const string UninstallID = "Zero Install_is1";
+
         /// <summary>
         /// Adds Zero Install to the start menu and the PATH environment variable.
         /// </summary>
-        private void DesktopIntegrationApply()
+        /// <param name="size">The size of the installed files in bytes.</param>
+        private void DesktopIntegrationApply(long size)
         {
+            if (!WindowsUtils.IsWindows) return;
+
             Handler.RunTask(new SimpleTask(Resources.DesktopIntegrationApply, () =>
             {
+                UninstallEntry.Register(
+                    UninstallID,
+                    new[] { Path.Combine(TargetDir, "0install-win.exe"), Self.Name, Self.Remove.Name },
+                    "Zero Install",
+                    new("https://0install.net/"),
+                    iconPath: Path.Combine(TargetDir, "ZeroInstall.exe"),
+                    AppInfo.Current.Version,
+                    size,
+                    MachineWide);
+
+                RegistryUtils.SetSoftwareString(@"Microsoft\PackageManagement", "ZeroInstall", Path.Combine(TargetDir, "ZeroInstall.OneGet.dll"), MachineWide);
+
+                PathEnv.AddDir(TargetDir, MachineWide);
+
                 Shortcut.Create(
                     path: Shortcut.GetStartMenuPath("", "Zero Install", MachineWide),
                     targetPath: Path.Combine(TargetDir, "ZeroInstall.exe"),
                     appId: "ZeroInstall");
-
-                PathEnv.AddDir(TargetDir, MachineWide);
             }));
         }
 
@@ -31,18 +50,18 @@ namespace ZeroInstall.Commands.Desktop.SelfManagement
         /// </summary>
         private void DesktopIntegrationRemove()
         {
-            static void DeleteIfExists(string path)
-            {
-                if (File.Exists(path)) File.Delete(path);
-            }
+            if (!WindowsUtils.IsWindows) return;
 
             Handler.RunTask(new SimpleTask(Resources.DesktopIntegrationRemove, () =>
             {
-                DeleteIfExists(Shortcut.GetStartMenuPath("", "Zero Install", MachineWide));
-                DeleteIfExists(Shortcut.GetStartMenuPath("Zero install", "Zero Install", MachineWide));
-                DeleteIfExists(Shortcut.GetDesktopPath("Zero Install", MachineWide));
+                string path = Shortcut.GetStartMenuPath("", "Zero Install", MachineWide);
+                if (File.Exists(path)) File.Delete(path);
 
                 PathEnv.RemoveDir(TargetDir, MachineWide);
+
+                RegistryUtils.DeleteSoftwareValue(@"Microsoft\PackageManagement", "ZeroInstall", MachineWide);
+
+                UninstallEntry.Unregister(UninstallID, MachineWide);
             }));
         }
     }
